Refactor component structure and remove unused assets
- Updated import paths for `HeroSection` and `Navbar` components.
- Deleted unused SVG assets: `menu-icon.svg` and `profile-icon.svg`.
- Removed several components including `BackgroundPageContainer`, `HeroText`, `SearchInput`, `HeroIconList`, `SearchField`, and `WaveUnderline` to streamline the codebase.
- Introduced a new structure for image assets in `images.ts` to utilize icons from `lucide-react` instead of SVG files.
