+++ src/components/common/search-input.tsx
