+++ src/components/hero-section/hero-icon-list.tsx
