+++ src/data/images.ts
@@ -1,10 +1,13 @@
-export const IMAGES = {
-  LOGO: {
-    ICON: "/ps-logo.svg",
-    TEXT: "/ps-logo-text.svg"
+import { ImageItem } from "@/src/types";
+import { MenuIcon, UserIcon } from "lucide-react";
+
+export const images: ImageItem = {
+  logo: {
+    icon: "/ps-logo.svg",
+    text: "/ps-logo-text.svg"
   },
-  ICONS: {
-    MENU: "/menu-icon.svg",
-    PROFILE: "/profile-icon.svg"
+  icons: {
+    menu: MenuIcon,
+    profile: UserIcon
   }
 } as const;
